requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
cssselect2==0.6.0
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
django-widget-tweaks==1.4.12
flake8==3.8.4
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
lxml==4.9.0
mccabe==0.6.1
oscrypto==1.3.0
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.6.0
pycparser==2.21
pyflakes==2.2.0
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
python-bidi==0.4.2
python-decouple==3.6
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.10
requests==2.28.1
six==1.16.0
sqlparse==0.3.1
svglib==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.8
